+++ twitchio/errors.py
@@ -69,10 +69,12 @@ class HTTPException(TwitchIOException):
     def __init__(
         self, message: str, *, reason: Optional[str] = None, status: Optional[int] = None, extra: Optional[Any] = None
     ):
-        self.message = message
+        self.message = f"{status}: {message}: {reason}"
         self.reason = reason
         self.status = status
         self.extra = extra
+
+        super().__init__(self.message)
 
 
 class Unauthorized(HTTPException):